+++ src/vivarium_public_health/population/data_transformations.py
@@ -54,21 +54,21 @@ def assign_demographic_proportions(
 
     population_data["P(sex, location, age| year)"] = (
         population_data.groupby("year_start", as_index=False)
-        .apply(lambda sub_pop: sub_pop[["value"]] / sub_pop["value"].sum())
+        .apply(lambda sub_pop: sub_pop[["value"]] / sub_pop["value"].sum(), include_groups=False)
         .reset_index(level=0)["value"]
         .fillna(0.0)
     )
 
     population_data["P(sex, location | age, year)"] = (
         population_data.groupby(["age", "year_start"], as_index=False)
-        .apply(lambda sub_pop: sub_pop[["value"]] / sub_pop["value"].sum())
+        .apply(lambda sub_pop: sub_pop[["value"]] / sub_pop["value"].sum(), include_groups=False)
         .reset_index(level=0)["value"]
         .fillna(0.0)
     )
 
     population_data["P(age | year, sex, location)"] = (
         population_data.groupby(["year_start", "sex", "location"], as_index=False)
-        .apply(lambda sub_pop: sub_pop[["value"]] / sub_pop["value"].sum())
+        .apply(lambda sub_pop: sub_pop[["value"]] / sub_pop["value"].sum(), include_groups=False)
         .reset_index(level=0)["value"]
         .fillna(0.0)
     )